Make utils an installable python package
