+++ presidential-elections/setup.py
@@ -1,0 +1,9 @@
+from setuptools import find_packages, setup
+
+setup(
+    name='utils',
+    packages=find_packages(),
+    version='0.1.0',
+    description='Utilities for presidential models',
+    author='Alexandre Andorra & Remi Louf',
+)
